[#70276] arm: Fix access caching if page doesn't have uniform attribution
Security attribution faults are detected for each access unless the
access on the same page was successful before in the same CPU privilege
and security context. Then we don't check security attribution again
unless TLB cache gets flushed, e.g. when SAU config changes.

However, security attribution isn't guaranteed to be the same for the
whole page as IDAU and SAU granularity is 32B and `TARGET_PAGE_SIZE` is
currently 1kB.

Therefore in every lookup we need to know whether the result applies to
the whole page because only then we can cache access permissions and
skip the security attribution check for future accesses to the same
page.